Bugs fixed and relevant files were updated.

+++ Data Mining Project.sql
@@ -5,8 +5,8 @@ CREATE TABLE `products` (
   `country_id` int,
   `condition_id` int,
   `product_name` varchar(255) UNIQUE,
-  `product_price` float,
-  `shipping_fee` float,
+  `product_price` decimal(10,3),
+  `shipping_fee` decimal(10,3),
   `page_number` int
 );
 
@@ -33,12 +33,12 @@ CREATE TABLE `conditions` (
 
 CREATE TABLE `currency` (
   `product_id` int,
-  `Israeli_Shekel_ILS` float,
-  `US_Dollar_USD` float,
-  `EU_Euro_EUR` float,
-  `GB_Pound_GBP` float,
-  `China_Yoan_CNY` float,
-  `Russia_Ruble_RUB` float
+  `Israeli_Shekel_ILS` decimal(10,3),
+  `US_Dollar_USD` decimal(10,3),
+  `EU_Euro_EUR` decimal(10,3),
+  `GB_Pound_GBP` decimal(10,3),
+  `China_Yoan_CNY` decimal(10,3),
+  `Russia_Ruble_RUB` decimal(10,3)
 );
 
 ALTER TABLE `products` ADD FOREIGN KEY (`seller_id`) REFERENCES `sellers` (`seller_id`);
